Move timezone utility to libraries
